+++ src/CHANGELOG.md
@@ -1,6 +1,9 @@
 **Changelog**
 
 v.3.11.1
+* [bug] Rewrite rules caused fatal error on some [environments](https://wordpress.org/support/topic/critical-error-575/ "Lumiere WordPress support") 
+* [bug] Version of template item wasn't updated, so the template kept using the removed trait frontend
+* [technical] Removed French po/mo since all French strings are now managed online
 
 v.3.11
 * [feature] Popups are fully compatible now with Polylang. Both taxonomy and popups are now fully translated into French.